detectpill/requirements.txt
ultralytics
rembg
boto3
tensorflow-cpu==2.9.1
numpy==1.26.1
pillow==10.1.0
